fix: resolver warnings de React Refresh separando componentes y utilidades
- Movidos hooks useInstrumentLimit y useCanAddInstrument a archivo separado
- Movidas badgeVariants y buttonVariants a archivos separados
- Agregado tsconfig.json raíz para compatibilidad con ESLint
- Los archivos de componentes ahora solo exportan componentes React
- Permite que React Fast Refresh funcione correctamente sin warnings
