+++ tsconfig.json
@@ -1,0 +1,41 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "useDefineForClassFields": true,
+    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "module": "ESNext",
+    "skipLibCheck": true,
+    "moduleResolution": "bundler",
+    "allowImportingTsExtensions": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "jsx": "react-jsx",
+    "strict": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "noFallthroughCasesInSwitch": true,
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["./frontend/src/*"],
+      "@/components/*": ["./frontend/src/components/*"],
+      "@/pages/*": ["./frontend/src/pages/*"],
+      "@/hooks/*": ["./frontend/src/hooks/*"],
+      "@/store/*": ["./frontend/src/store/*"],
+      "@/utils/*": ["./frontend/src/utils/*"],
+      "@cedears-manager/shared/types": ["./shared/src/types/index"]
+    }
+  },
+  "include": [
+    "frontend/src/**/*",
+    "backend/src/**/*",
+    "shared/src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist",
+    "build",
+    "**/dist/**",
+    "**/build/**"
+  ]
+}
